--- Landmarks/ContentView.swift
@@ -1,31 +1,0 @@
-//
-//  ContentView.swift
-//  Landmarks
-//
-//  Created by Rafael Ferreira on 2/11/21.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack(alignment: .leading) {
-            Text("Turtle Rock")
-                .font(.title)
-            HStack {
-                Text("Joshua Tree National Park")
-                    .font(.subheadline)
-                Spacer()
-                Text("California")
-                    .font(.subheadline)
-            }
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
