+++ Landmarks/DetailView.swift
@@ -1,0 +1,50 @@
+//
+//  DetailView.swift
+//  Landmarks
+//
+//  Created by Rafael Ferreira on 2/11/21.
+//
+
+import SwiftUI
+
+struct DetailView: View {
+    var body: some View {
+        VStack {
+            MapView()
+                .ignoresSafeArea(edges: .top)
+                .frame(height: 300)
+
+            CircleImage()
+                .offset(y: -130)
+                .padding(.bottom, -130)
+
+            VStack(alignment: .leading) {
+                Text("Turtle Rock")
+                    .font(.title)
+
+                HStack {
+                    Text("Joshua Tree National Park")
+                    Spacer()
+                    Text("California")
+                }
+                .font(.subheadline)
+
+                Divider()
+
+                Text("About Turtle Rock")
+                    .font(.title2)
+
+                Text("Descriptive text goes here.")
+            }
+            .padding()
+
+            Spacer()
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        DetailView()
+    }
+}
