+++ Landmarks/MapView.swift
@@ -1,0 +1,26 @@
+//
+//  MapView.swift
+//  Landmarks
+//
+//  Created by Rafael Ferreira on 2/11/21.
+//
+
+import MapKit
+import SwiftUI
+
+struct MapView: View {
+    @State private var region = MKCoordinateRegion(
+        center: CLLocationCoordinate2D(latitude: 34.011_286, longitude: -116.166_868),
+        span: MKCoordinateSpan(latitudeDelta: 0.2, longitudeDelta: 0.2)
+    )
+
+    var body: some View {
+        Map(coordinateRegion: $region)
+    }
+}
+
+struct MapView_Previews: PreviewProvider {
+    static var previews: some View {
+        MapView()
+    }
+}
